{"schema":"decodex/commit/1","summary":"Split config and Codex account modules","authority":"manual"}
