+++ apps/decodex/src/agent/codex_accounts/pool.rs
@@ -1,0 +1,629 @@
+use std::{
+	path::{Path, PathBuf},
+	sync::Mutex,
+};
+
+use color_eyre::Report;
+use reqwest::{StatusCode, blocking::Client};
+use serde_json::Value;
+use time::{OffsetDateTime, format_description::well_known::Rfc3339};
+
+use crate::{
+	agent::codex_accounts::{
+		CHATGPT_OAUTH_CLIENT_ID, CODEX_USER_AGENT, CodexAccountAuthFailure, CodexAccountLogin,
+		DEFAULT_REFRESH_ENDPOINT, DEFAULT_USAGE_ENDPOINT, HTTP_TIMEOUT,
+		record::{self, AccountPoolRecord},
+		refresh::{
+			self, ProactiveRefreshError, RefreshRequest, RefreshResponse, RefreshStatus,
+			ReportableRefreshError,
+		},
+		selection,
+		usage::{self, AccountProfileSnapshot, AccountUsageSnapshot, UsageProbeError},
+	},
+	config::ProjectCodexAccountsConfig,
+	prelude::eyre,
+	runtime,
+};
+
+pub(crate) trait CodexAccountProvider {
+	fn select_account(&self) -> crate::prelude::Result<CodexAccountLogin>;
+	fn refresh_account(
+		&self,
+		previous_account_id: Option<&str>,
+	) -> crate::prelude::Result<CodexAccountLogin>;
+}
+
+pub(crate) struct CodexAccountPool {
+	pub(super) path: PathBuf,
+	pub(super) usage_endpoint: String,
+	pub(super) profile_endpoint: Option<String>,
+	pub(super) refresh_endpoint: String,
+	pub(super) fixed_account: Option<String>,
+	pub(super) codex_auth_path: PathBuf,
+	pub(super) client: Client,
+	pub(super) selected_account_id: Mutex<Option<String>>,
+}
+impl CodexAccountPool {
+	pub(crate) fn from_config(config: &ProjectCodexAccountsConfig) -> crate::prelude::Result<Self> {
+		let fixed_account = runtime::global_fixed_account_selector()?;
+		let usage_endpoint = config.usage_endpoint().unwrap_or(DEFAULT_USAGE_ENDPOINT);
+
+		Self::new_with_fixed_account_and_profile_endpoint(
+			runtime::accounts_path()?,
+			usage_endpoint,
+			config.profile_endpoint(),
+			config.refresh_endpoint().unwrap_or(DEFAULT_REFRESH_ENDPOINT),
+			fixed_account.as_deref(),
+		)
+	}
+
+	pub(crate) fn from_accounts_path(path: impl AsRef<Path>) -> crate::prelude::Result<Self> {
+		Self::new_with_fixed_account(path, DEFAULT_USAGE_ENDPOINT, DEFAULT_REFRESH_ENDPOINT, None)
+	}
+
+	pub(super) fn new_with_fixed_account(
+		path: impl AsRef<Path>,
+		usage_endpoint: impl Into<String>,
+		refresh_endpoint: impl Into<String>,
+		fixed_account: Option<&str>,
+	) -> crate::prelude::Result<Self> {
+		Self::new_with_fixed_account_and_profile_endpoint(
+			path,
+			usage_endpoint,
+			None,
+			refresh_endpoint,
+			fixed_account,
+		)
+	}
+
+	fn new_with_fixed_account_and_profile_endpoint(
+		path: impl AsRef<Path>,
+		usage_endpoint: impl Into<String>,
+		profile_endpoint: Option<&str>,
+		refresh_endpoint: impl Into<String>,
+		fixed_account: Option<&str>,
+	) -> crate::prelude::Result<Self> {
+		Self::new_with_fixed_account_profile_and_codex_auth_path(
+			path,
+			usage_endpoint,
+			profile_endpoint,
+			refresh_endpoint,
+			fixed_account,
+			record::default_codex_auth_json_path()?,
+		)
+	}
+
+	#[cfg(test)]
+	pub(super) fn new_with_fixed_account_and_codex_auth_path(
+		path: impl AsRef<Path>,
+		usage_endpoint: impl Into<String>,
+		refresh_endpoint: impl Into<String>,
+		fixed_account: Option<&str>,
+		codex_auth_path: impl Into<PathBuf>,
+	) -> crate::prelude::Result<Self> {
+		Self::new_with_fixed_account_profile_and_codex_auth_path(
+			path,
+			usage_endpoint,
+			None,
+			refresh_endpoint,
+			fixed_account,
+			codex_auth_path,
+		)
+	}
+
+	fn new_with_fixed_account_profile_and_codex_auth_path(
+		path: impl AsRef<Path>,
+		usage_endpoint: impl Into<String>,
+		profile_endpoint: Option<&str>,
+		refresh_endpoint: impl Into<String>,
+		fixed_account: Option<&str>,
+		codex_auth_path: impl Into<PathBuf>,
+	) -> crate::prelude::Result<Self> {
+		let client = Client::builder().timeout(HTTP_TIMEOUT).build()?;
+		let usage_endpoint = usage_endpoint.into();
+		let profile_endpoint = profile_endpoint
+			.and_then(|endpoint| usage::nonblank_string(Some(endpoint)))
+			.or_else(|| record::default_profile_endpoint_for_usage_endpoint(&usage_endpoint));
+
+		Ok(Self {
+			path: path.as_ref().to_path_buf(),
+			usage_endpoint,
+			profile_endpoint,
+			refresh_endpoint: refresh_endpoint.into(),
+			fixed_account: fixed_account
+				.map(str::trim)
+				.filter(|selector| !selector.is_empty())
+				.map(str::to_owned),
+			codex_auth_path: codex_auth_path.into(),
+			client,
+			selected_account_id: Mutex::new(None),
+		})
+	}
+
+	fn select_from_records(
+		&self,
+		records: &mut [AccountPoolRecord],
+	) -> crate::prelude::Result<CodexAccountLogin> {
+		if let Some(selector) = self.fixed_account.as_deref() {
+			return self.select_fixed_from_records(records, selector);
+		}
+
+		let now = OffsetDateTime::now_utc().unix_timestamp();
+		let mut candidates = Vec::new();
+		let mut skipped = Vec::new();
+		let mut records_changed = false;
+
+		for (index, record) in records.iter_mut().enumerate() {
+			match self.account_candidate_from_record(
+				record,
+				index + 1,
+				now,
+				&mut records_changed,
+			)? {
+				Ok(candidate) => candidates.push(candidate),
+				Err(reason) => skipped.push(reason),
+			}
+		}
+
+		if records_changed {
+			self.save_records(records)?;
+		}
+		if candidates.is_empty() {
+			if let Some(auth_failure) =
+				records.iter().find_map(AccountPoolRecord::auth_failed_error)
+			{
+				return Err(Report::new(auth_failure));
+			}
+
+			eyre::bail!(
+				"No usable Codex account was available from `{}`. Skipped entries: {}",
+				self.path.display(),
+				if skipped.is_empty() { String::from("none") } else { skipped.join("; ") }
+			);
+		}
+
+		candidates.sort_by(selection::compare_account_candidates);
+
+		let mut selected = candidates.remove(0);
+
+		selected.mark_selected(now);
+
+		if let Some(record) =
+			records.iter_mut().find(|record| record.account_id() == Some(selected.account_id()))
+		{
+			record.last_selected_at_unix_epoch = Some(now);
+			records_changed = true;
+		}
+
+		let account_summaries = selection::account_summaries(&selected, &candidates);
+		let selected = selected.with_account_summaries(account_summaries);
+
+		if records_changed {
+			self.save_records(records)?;
+		}
+
+		self.remember_selected_account(&selected.account_id)?;
+
+		Ok(selected)
+	}
+
+	fn select_fixed_from_records(
+		&self,
+		records: &mut [AccountPoolRecord],
+		selector: &str,
+	) -> crate::prelude::Result<CodexAccountLogin> {
+		let record_index = self.fixed_record_index(records, selector)?;
+		let now = OffsetDateTime::now_utc().unix_timestamp();
+		let mut records_changed = false;
+
+		if let Some(auth_failure) = records[record_index].auth_failed_error() {
+			return Err(Report::new(auth_failure));
+		}
+
+		let mut selected = match self.account_candidate_from_record(
+			&mut records[record_index],
+			record_index + 1,
+			now,
+			&mut records_changed,
+		)? {
+			Ok(candidate) => candidate,
+			Err(reason) => {
+				eyre::bail!(
+					"Configured Codex fixed account `{selector}` from `{}` is not usable: {reason}",
+					self.path.display()
+				);
+			},
+		};
+
+		selected.mark_selected(now);
+
+		records[record_index].last_selected_at_unix_epoch = Some(now);
+		records_changed = true;
+
+		let account_summaries = selection::account_summaries(&selected, &[]);
+		let selected = selected.with_account_summaries(account_summaries);
+
+		if records_changed {
+			self.save_records(records)?;
+		}
+
+		self.remember_selected_account(&selected.account_id)?;
+
+		Ok(selected)
+	}
+
+	fn account_candidate_from_record(
+		&self,
+		record: &mut AccountPoolRecord,
+		line_number: usize,
+		now: i64,
+		records_changed: &mut bool,
+	) -> crate::prelude::Result<std::result::Result<CodexAccountLogin, String>> {
+		if record.disabled {
+			return Ok(Err(format!("line {line_number} disabled")));
+		}
+
+		if let Some(auth_failure) = record.auth_failure() {
+			return Ok(Err(format!("line {line_number} auth failed: {auth_failure}")));
+		}
+
+		if record.cooldown_until_unix_epoch.is_some_and(|cooldown| cooldown > now) {
+			return Ok(Err(format!("line {line_number} cooling down")));
+		}
+		if record.account_id().is_none() {
+			return Ok(Err(format!("line {line_number} missing account id")));
+		}
+		if record.access_token().is_none() {
+			return Ok(Err(format!("line {line_number} missing access token")));
+		}
+
+		let refresh_status = match self.proactive_refresh_record(record, now) {
+			Ok(status) => {
+				if status == RefreshStatus::Succeeded {
+					*records_changed = true;
+				}
+
+				status.as_str()
+			},
+			Err(error) if error.auth_failed => {
+				*records_changed = true;
+
+				return Ok(Err(format!("{} auth failed: {}", record.display_name(), error.source)));
+			},
+			Err(error) if error.requires_skip => {
+				return Ok(Err(format!(
+					"{} proactive refresh failed: {}",
+					record.display_name(),
+					error.source
+				)));
+			},
+			Err(_error) => RefreshStatus::Failed.as_str(),
+		};
+
+		match self.probe_record_usage(record) {
+			Ok(usage) => Ok(Ok(record.login_from_usage(usage, refresh_status)?)),
+			Err(error) if error.unauthorized && record.refresh_token().is_some() => {
+				if let Err(refresh_error) = self.refresh_record(record) {
+					if let Some(auth_failure) =
+						refresh_error.downcast_ref::<CodexAccountAuthFailure>()
+					{
+						*records_changed = true;
+
+						return Ok(Err(format!(
+							"{} auth failed: {auth_failure}",
+							record.display_name()
+						)));
+					}
+
+					return Err(refresh_error);
+				}
+
+				*records_changed = true;
+
+				let usage = self.probe_record_usage(record).map_err(|retry_error| {
+					eyre::eyre!(
+						"Codex account `{}` refreshed but usage probe still failed: {retry_error}",
+						record.display_name()
+					)
+				})?;
+
+				Ok(Ok(record.login_from_usage(usage, "succeeded")?))
+			},
+			Err(error) => Ok(Err(format!("{} usage probe failed: {error}", record.display_name()))),
+		}
+	}
+
+	pub(super) fn proactive_refresh_record(
+		&self,
+		record: &mut AccountPoolRecord,
+		now_unix_epoch: i64,
+	) -> std::result::Result<RefreshStatus, ProactiveRefreshError> {
+		let Some(reason) = record.proactive_refresh_reason(now_unix_epoch) else {
+			return Ok(RefreshStatus::NotNeeded);
+		};
+
+		if record.refresh_token().is_none() {
+			return Err(ProactiveRefreshError {
+				source: ReportableRefreshError::new(format!("missing refresh token for {reason}")),
+				requires_skip: reason.requires_valid_token(),
+				auth_failed: false,
+			});
+		}
+
+		self.refresh_record(record).map(|()| RefreshStatus::Succeeded).map_err(|error| {
+			let auth_failed = error.downcast_ref::<CodexAccountAuthFailure>().is_some();
+
+			ProactiveRefreshError {
+				source: ReportableRefreshError::new(error.to_string()),
+				requires_skip: reason.requires_valid_token(),
+				auth_failed,
+			}
+		})
+	}
+
+	fn refresh_from_records(
+		&self,
+		records: &mut [AccountPoolRecord],
+		previous_account_id: Option<&str>,
+	) -> crate::prelude::Result<CodexAccountLogin> {
+		let record_index = if let Some(selector) = self.fixed_account.as_deref() {
+			self.fixed_record_index(records, selector)?
+		} else {
+			let selected_account_id = self.selected_account_id()?;
+			let target_account_id = previous_account_id.or(selected_account_id.as_deref());
+
+			records
+				.iter()
+				.position(|record| {
+					target_account_id.is_none_or(|target| record.account_id() == Some(target))
+				})
+				.ok_or_else(|| {
+					eyre::eyre!(
+						"Codex account refresh requested an account that is not in the configured accounts."
+					)
+				})?
+		};
+
+		if let Some(auth_failure) = records[record_index].auth_failed_error() {
+			return Err(Report::new(auth_failure));
+		}
+		if let Err(error) = self.refresh_record(&mut records[record_index]) {
+			if error.downcast_ref::<CodexAccountAuthFailure>().is_some() {
+				self.save_records(records)?;
+			}
+
+			return Err(error);
+		}
+
+		let usage = self.probe_record_usage(&records[record_index])?;
+		let now = OffsetDateTime::now_utc().unix_timestamp();
+		let mut selected = records[record_index].login_from_usage(usage, "succeeded")?;
+
+		selected.mark_selected(now);
+
+		records[record_index].last_selected_at_unix_epoch = Some(now);
+
+		let selected_summary = selected.summary().clone();
+		let selected = selected.with_account_summaries(vec![selected_summary]);
+
+		self.save_records(records)?;
+		self.remember_selected_account(&selected.account_id)?;
+
+		Ok(selected)
+	}
+
+	fn fixed_record_index(
+		&self,
+		records: &[AccountPoolRecord],
+		selector: &str,
+	) -> crate::prelude::Result<usize> {
+		let matches = records
+			.iter()
+			.enumerate()
+			.filter_map(|(index, record)| {
+				record.matches_account_selector(selector).then_some(index)
+			})
+			.collect::<Vec<_>>();
+
+		match matches.as_slice() {
+			[] => eyre::bail!(
+				"Configured Codex fixed account `{selector}` does not match any account in `{}`.",
+				self.path.display()
+			),
+			[index] => Ok(*index),
+			_ => eyre::bail!(
+				"Configured Codex fixed account `{selector}` matched multiple accounts in `{}`.",
+				self.path.display()
+			),
+		}
+	}
+
+	pub(super) fn probe_record_usage(
+		&self,
+		record: &AccountPoolRecord,
+	) -> std::result::Result<AccountUsageSnapshot, UsageProbeError> {
+		let access_token = record
+			.access_token()
+			.ok_or_else(|| UsageProbeError::other("account is missing an access token"))?;
+		let account_id = record
+			.account_id()
+			.ok_or_else(|| UsageProbeError::other("account is missing an account id"))?;
+		let response = self
+			.client
+			.get(&self.usage_endpoint)
+			.bearer_auth(access_token)
+			.header("ChatGPT-Account-Id", account_id)
+			.header("User-Agent", CODEX_USER_AGENT)
+			.send()
+			.map_err(|error| UsageProbeError::other(error.to_string()))?;
+		let status = response.status();
+
+		if status == StatusCode::UNAUTHORIZED {
+			return Err(UsageProbeError::unauthorized());
+		}
+		if !status.is_success() {
+			return Err(UsageProbeError::other(format!("usage endpoint returned {status}")));
+		}
+
+		let payload = response.json::<Value>().map_err(|error| {
+			UsageProbeError::other(format!("usage JSON did not parse: {error}"))
+		})?;
+
+		Ok(usage::usage_snapshot_from_payload(&payload, OffsetDateTime::now_utc().unix_timestamp()))
+	}
+
+	pub(super) fn probe_record_profile(
+		&self,
+		record: &AccountPoolRecord,
+	) -> std::result::Result<Option<AccountProfileSnapshot>, UsageProbeError> {
+		let Some(profile_endpoint) = self.profile_endpoint.as_deref() else {
+			return Ok(None);
+		};
+		let access_token = record
+			.access_token()
+			.ok_or_else(|| UsageProbeError::other("account is missing an access token"))?;
+		let account_id = record
+			.account_id()
+			.ok_or_else(|| UsageProbeError::other("account is missing an account id"))?;
+		let response = self
+			.client
+			.get(profile_endpoint)
+			.bearer_auth(access_token)
+			.header("ChatGPT-Account-Id", account_id)
+			.header("User-Agent", CODEX_USER_AGENT)
+			.send()
+			.map_err(|error| UsageProbeError::other(error.to_string()))?;
+		let status = response.status();
+
+		if status == StatusCode::UNAUTHORIZED {
+			return Err(UsageProbeError::unauthorized());
+		}
+		if !status.is_success() {
+			return Err(UsageProbeError::other(format!("profile endpoint returned {status}")));
+		}
+
+		let payload = response.json::<Value>().map_err(|error| {
+			UsageProbeError::other(format!("profile JSON did not parse: {error}"))
+		})?;
+
+		Ok(usage::profile_snapshot_from_payload(
+			&payload,
+			OffsetDateTime::now_utc().unix_timestamp(),
+		))
+	}
+
+	pub(super) fn refresh_record(
+		&self,
+		record: &mut AccountPoolRecord,
+	) -> crate::prelude::Result<()> {
+		let display_name = record.display_name();
+		let refresh_token = record
+			.refresh_token()
+			.ok_or_else(|| {
+				eyre::eyre!(
+					"Codex account `{}` cannot refresh because no refresh token is present.",
+					display_name
+				)
+			})?
+			.to_owned();
+		let response = self
+			.client
+			.post(&self.refresh_endpoint)
+			.header("Content-Type", "application/json")
+			.json(&RefreshRequest {
+				client_id: CHATGPT_OAUTH_CLIENT_ID,
+				grant_type: "refresh_token",
+				refresh_token,
+			})
+			.send()?;
+		let status = response.status();
+
+		if !status.is_success() {
+			let reason = format!(
+				"Codex account `{}` token refresh failed with HTTP {status}.",
+				display_name
+			);
+
+			if refresh::token_refresh_auth_status(status) {
+				record.mark_auth_failed(OffsetDateTime::now_utc().unix_timestamp(), reason.clone());
+
+				return Err(Report::new(CodexAccountAuthFailure::from_record(record, reason)));
+			}
+
+			eyre::bail!("{reason}");
+		}
+
+		let refresh_response = response.json::<RefreshResponse>()?;
+		let tokens = record.tokens.as_mut().ok_or_else(|| {
+			eyre::eyre!("Codex account `{display_name}` is missing token storage.")
+		})?;
+
+		if let Some(id_token) = refresh_response.id_token {
+			tokens.id_token = Some(id_token);
+		}
+		if let Some(access_token) = refresh_response.access_token {
+			tokens.access_token = access_token;
+		}
+		if let Some(refresh_token) = refresh_response.refresh_token {
+			tokens.refresh_token = refresh_token;
+		}
+
+		if tokens.access_token.trim().is_empty() {
+			eyre::bail!(
+				"Codex account `{}` token refresh did not produce a usable access token.",
+				display_name
+			);
+		}
+
+		record.last_refresh = Some(OffsetDateTime::now_utc().format(&Rfc3339)?);
+
+		record.clear_auth_failed();
+		self.sync_codex_auth_for_refreshed_record(record)?;
+
+		Ok(())
+	}
+
+	fn sync_codex_auth_for_refreshed_record(
+		&self,
+		record: &AccountPoolRecord,
+	) -> crate::prelude::Result<()> {
+		record::sync_refreshed_record_to_codex_auth(record, &self.codex_auth_path)
+	}
+
+	fn remember_selected_account(&self, account_id: &str) -> crate::prelude::Result<()> {
+		let mut selected = self
+			.selected_account_id
+			.lock()
+			.map_err(|_| eyre::eyre!("Codex accounts selection lock was poisoned."))?;
+
+		*selected = Some(account_id.to_owned());
+
+		Ok(())
+	}
+
+	fn selected_account_id(&self) -> crate::prelude::Result<Option<String>> {
+		self.selected_account_id
+			.lock()
+			.map(|selected| selected.clone())
+			.map_err(|_| eyre::eyre!("Codex accounts selection lock was poisoned."))
+	}
+}
+
+impl CodexAccountProvider for CodexAccountPool {
+	fn select_account(&self) -> crate::prelude::Result<CodexAccountLogin> {
+		let _guard = self.lock_records()?;
+		let mut records = self.load_records()?;
+
+		self.select_from_records(&mut records)
+	}
+
+	fn refresh_account(
+		&self,
+		previous_account_id: Option<&str>,
+	) -> crate::prelude::Result<CodexAccountLogin> {
+		let _guard = self.lock_records()?;
+		let mut records = self.load_records()?;
+
+		self.refresh_from_records(&mut records, previous_account_id)
+	}
+}
